requirements.txt
torch
"transformers[torch]"
huggingface_hub
python-dotenv
"pymongo[srv]"
streamlit
